sqon removal
